refactor(util/user): Move isHome and isWork to util.
