+++ lib/components/user/favorite-locations-pane.js
@@ -2,7 +2,8 @@ import { FieldArray } from 'formik'
 import React from 'react'
 import { ControlLabel } from 'react-bootstrap'
 
-import FavoriteLocation, { isHome, isWork } from './favorite-location'
+import FavoriteLocation from './favorite-location'
+import { isHome, isWork } from '../../util/user'
 
 /**
  * Renders an editable list user's favorite locations, and lets the user add a new one.
