+++ lib/components/user/user-account-screen.js
@@ -8,11 +8,11 @@ import * as yup from 'yup'
 import * as uiActions from '../../actions/ui'
 import * as userActions from '../../actions/user'
 import { RETURN_TO_CURRENT_ROUTE } from '../../util/ui'
-import { isNewUser } from '../../util/user'
+import { isNewUser, isHome, isWork } from '../../util/user'
 import DesktopNav from '../app/desktop-nav'
 import AccountSetupFinishPane from './account-setup-finish-pane'
 import ExistingAccountDisplay from './existing-account-display'
-import { isHome, isWork, BLANK_HOME, BLANK_WORK } from './favorite-location'
+import { BLANK_HOME, BLANK_WORK } from './favorite-location'
 import FavoriteLocationsPane from './favorite-locations-pane'
 import NewAccountWizard from './new-account-wizard'
 import NotificationPrefsPane from './notification-prefs-pane'
